Implement appropriate methods

+++ LagerhotellAPI/Services/WarehouseHotelService.cs
@@ -15,6 +15,7 @@ public class WarehouseHotelService
     public async Task<string> AddWarehouseHotel(WarehouseHotel warehouseHotel)
     {
         string id = Guid.NewGuid().ToString();
+        // Should check if there exists one with the same name
         Models.DbModels.WarehouseHotel dbWarehouseHotel = new(id, warehouseHotel.Coordinate, warehouseHotel.Address, warehouseHotel.Name, warehouseHotel.OpeningHours, warehouseHotel.PhoneNumber, warehouseHotel.DetailedDescription, warehouseHotel.ContainsTemperatedStorageUnits, warehouseHotel.IsActive, warehouseHotel.StorageUnitsSizes);
         await _warehouseHotels.InsertOneAsync(dbWarehouseHotel);
         return id;
@@ -23,23 +24,42 @@ public class WarehouseHotelService
     public async Task<(string, string)> DeleteWarehouseHotel(string id)
     {
         WarehouseHotel deletedWarehouseHotel = await GetWarehouseHotelById(id);
+        if (deletedWarehouseHotel == null)
+        {
+            throw new KeyNotFoundException();
+        }
         await _warehouseHotels.DeleteOneAsync(hotel => hotel.WarehouseHotelId == id);
         return (deletedWarehouseHotel.WarehouseHotelId, deletedWarehouseHotel.Name);
     }
 
-    public async Task ModifyWarehouseHotel(WarehouseHotel warehouseHotel)
+    public async Task ModifyWarehouseHotel(WarehouseHotel warehouseHotel, string oldName)
     {
         Models.DbModels.WarehouseHotel oldWarehouseHotel = await GetWarehouseHotelByIdDbModel(warehouseHotel.WarehouseHotelId);
         if (oldWarehouseHotel != null)
         {
             LagerhotellAPI.Models.DbModels.WarehouseHotel dbWarehouseHotel = new(oldWarehouseHotel.Id, warehouseHotel.WarehouseHotelId, warehouseHotel.Coordinate, warehouseHotel.Address, warehouseHotel.Name, warehouseHotel.OpeningHours, warehouseHotel.PhoneNumber, warehouseHotel.DetailedDescription, warehouseHotel.ContainsTemperatedStorageUnits, warehouseHotel.IsActive, warehouseHotel.StorageUnitsSizes);
-            await _warehouseHotels.ReplaceOneAsync(hotel => hotel.WarehouseHotelId == warehouseHotel.WarehouseHotelId, dbWarehouseHotel);
+            await _warehouseHotels.ReplaceOneAsync(hotel => hotel.Name == oldName, dbWarehouseHotel);
+        }
+        else
+        {
+            throw new KeyNotFoundException();
         }
     }
 
     public async Task<WarehouseHotel> GetWarehouseHotelById(string id)
     {
         LagerhotellAPI.Models.DbModels.WarehouseHotel dbWarehouseHotel = await _warehouseHotels.Find(hotel => hotel.WarehouseHotelId == id).FirstOrDefaultAsync();
+        if (dbWarehouseHotel == null || dbWarehouseHotel.WarehouseHotelId == null)
+        {
+            throw new KeyNotFoundException();
+        }
+        WarehouseHotel domainWarehouseHotel = new(dbWarehouseHotel.WarehouseHotelId, dbWarehouseHotel.Coordinate, dbWarehouseHotel.Address, dbWarehouseHotel.Name, dbWarehouseHotel.OpeningHours, dbWarehouseHotel.PhoneNumber, dbWarehouseHotel.DetailedDescription, dbWarehouseHotel.ContainsTemperatedStorageUnits, dbWarehouseHotel.IsActive, dbWarehouseHotel.StorageUnitsSizes);
+        return domainWarehouseHotel;
+    }
+
+    public async Task<WarehouseHotel> GetWarehouseHotelByName(string name)
+    {
+        LagerhotellAPI.Models.DbModels.WarehouseHotel dbWarehouseHotel = await _warehouseHotels.Find(hotel => hotel.Name == name).FirstOrDefaultAsync();
         if (dbWarehouseHotel == null || dbWarehouseHotel.WarehouseHotelId == null)
         {
             throw new KeyNotFoundException();
